changed sanitize to sanitize json bodies instead of individual strings
