+++ usaspending_api/search/v2/elasticsearch_helper.py
@@ -16,26 +16,6 @@ KEYWORD_DATATYPE_FIELDS = ["{}.raw".format(i) for i in KEYWORD_DATATYPE_FIELDS]
 TRANSACTIONS_LOOKUP.update({v: k for k, v in TRANSACTIONS_LOOKUP.items()})
 
 
-def preprocess(keyword):
-    keyword = concat_if_array(keyword)
-    """Remove Lucene special characters instead of escaping for now"""
-    processed_string = re.sub(r"[/:][^!]", "", keyword)
-    if len(processed_string) != len(keyword):
-        msg = "Stripped characters from ES keyword search string New: '{}' Original: '{}'"
-        logger.info(msg.format(processed_string, keyword))
-        keyword = processed_string
-    return keyword
-
-
-def es_sanitize(keyword):
-    """ Escapes reserved elasticsearch characters and removes when necessary """
-    processed_string = re.sub(r'([-&!|{}()^~*?:\\/"+\[\]<>])', '', keyword)
-    if len(processed_string) != len(keyword):
-        msg = "Stripped characters from ES search string New: '{}' Original: '{}'"
-        logger.info(msg.format(processed_string, keyword))
-    return processed_string
-
-
 def swap_keys(dictionary_):
     return dict((TRANSACTIONS_LOOKUP.get(old_key, old_key), new_key) for (old_key, new_key) in dictionary_.items())
